+++ src/Admin/BankAdmin.php
@@ -101,7 +101,6 @@ class BankAdmin extends AbstractBaseAdmin
      */
     protected function configureListFields(ListMapper $listMapper): void
     {
-        unset($this->listModes['mosaic']);
         $listMapper
             ->add(
                 'name',