+++ src/Admin/BlogCategoryAdmin.php
@@ -1,0 +1,201 @@
+<?php
+
+namespace App\Admin;
+
+use Doctrine\ORM\QueryBuilder;
+use Sonata\AdminBundle\Datagrid\ListMapper;
+use Sonata\AdminBundle\Datagrid\DatagridMapper;
+use Sonata\AdminBundle\Form\FormMapper;
+use Sonata\AdminBundle\Route\RouteCollection;
+use Sonata\AdminBundle\Show\ShowMapper;
+use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
+
+/**
+ * Class BlogCategoryAdmin.
+ *
+ * @category Admin
+ */
+class BlogCategoryAdmin extends AbstractBaseAdmin
+{
+    protected $classnameLabel = 'Categoria';
+    protected $baseRoutePattern = 'web/blog-category';
+    protected $datagridValues = array(
+        '_sort_by' => 'title',
+        '_sort_order' => 'asc',
+    );
+
+    /**
+     * Configure route collection.
+     *
+     * @param RouteCollection $collection
+     */
+    protected function configureRoutes(RouteCollection $collection)
+    {
+        $collection->remove('batch');
+    }
+
+    /**
+     * Override query list to reduce queries amount on list view (apply join strategy).
+     *
+     * @param string $context context
+     *
+     * @return QueryBuilder
+     */
+    public function createQuery($context = 'list')
+    {
+        /** @var QueryBuilder $query */
+        $query = parent::createQuery($context);
+        $query
+            ->select($query->getRootAliases()[0].', p')
+            ->leftJoin($query->getRootAliases()[0].'.posts', 'p');
+
+        return $query;
+    }
+
+    /**
+     * @param FormMapper $formMapper
+     */
+    protected function configureFormFields(FormMapper $formMapper)
+    {
+        $formMapper
+            ->with('backend.admin.category', $this->getFormMdSuccessBoxArray(6))
+            ->add(
+                'title',
+                null,
+                array(
+                    'label' => 'backend.admin.title',
+                )
+            )
+            ->end()
+            ->with('backend.admin.controls', $this->getFormMdSuccessBoxArray(6))
+            ->add(
+                'enabled',
+                CheckboxType::class,
+                array(
+                    'label' => 'backend.admin.enabled',
+                    'required' => false,
+                )
+            )
+            ->end()
+        ;
+    }
+
+    /**
+     * @param DatagridMapper $datagridMapper
+     */
+    protected function configureDatagridFilters(DatagridMapper $datagridMapper)
+    {
+        $datagridMapper
+            ->add(
+                'title',
+                null,
+                array(
+                    'label' => 'backend.admin.title',
+                )
+            )
+            ->add(
+                'posts',
+                null,
+                array(
+                    'label' => 'backend.admin.posts',
+                )
+            )
+            ->add(
+                'enabled',
+                null,
+                array(
+                    'label' => 'backend.admin.enabled',
+                )
+            )
+        ;
+    }
+
+    /**
+     * @param ShowMapper $showMapper
+     */
+    protected function configureShowFields(ShowMapper $showMapper)
+    {
+        $showMapper
+            ->add(
+                'createdAt',
+                'date',
+                array(
+                    'label' => 'backend.admin.date',
+                    'format' => 'd/m/Y H:i',
+                )
+            )
+            ->add(
+                'title',
+                null,
+                array(
+                    'label' => 'backend.admin.title',
+                )
+            )
+            ->add(
+                'posts',
+                null,
+                array(
+                    'label' => 'backend.admin.posts',
+                )
+            )
+            ->add(
+                'enabled',
+                null,
+                array(
+                    'label' => 'backend.admin.enabled',
+                )
+            )
+        ;
+    }
+
+    /**
+     * @param ListMapper $listMapper
+     */
+    protected function configureListFields(ListMapper $listMapper)
+    {
+        $listMapper
+            ->add(
+                'title',
+                null,
+                array(
+                    'label' => 'backend.admin.title',
+                    'editable' => true,
+                )
+            )
+            ->add(
+                'count',
+                null,
+                array(
+                    'label' => 'backend.admin.posts_amount',
+                    'template' => 'Admin/Cells/list__cell_posts_amount_field.html.twig',
+                )
+            )
+            ->add(
+                'enabled',
+                null,
+                array(
+                    'label' => 'backend.admin.enabled',
+                    'editable' => true,
+                )
+            )
+            ->add(
+                '_action',
+                'actions',
+                array(
+                    'actions' => array(
+                        'show' => array(
+                            'template' => 'Admin/Buttons/list__action_show_button.html.twig',
+                        ),
+                        'edit' => array(
+                            'template' => 'Admin/Buttons/list__action_edit_button.html.twig',
+                        ),
+                        'delete' => array(
+                            'template' => 'Admin/Buttons/list__action_delete_button.html.twig',
+                        ),
+                    ),
+                    'label' => 'backend.admin.actions',
+                )
+            )
+        ;
+    }
+}