+++ src/Admin/CityAdmin.php
@@ -123,7 +123,6 @@ class CityAdmin extends AbstractBaseAdmin
      */
     protected function configureListFields(ListMapper $listMapper): void
     {
-        unset($this->listModes['mosaic']);
         $listMapper
             ->add(
                 'postalCode',