+++ src/Admin/ClassGroupAdmin.php
@@ -146,7 +146,6 @@ class ClassGroupAdmin extends AbstractBaseAdmin
      */
     protected function configureListFields(ListMapper $listMapper): void
     {
-        unset($this->listModes['mosaic']);
         $listMapper
             ->add(
                 'code',