+++ src/Admin/ContactMessageAdmin.php
@@ -173,7 +173,6 @@ class ContactMessageAdmin extends AbstractBaseAdmin
      */
     protected function configureListFields(ListMapper $listMapper): void
     {
-        unset($this->listModes['mosaic']);
         $listMapper
             ->add(
                 'checked',