+++ src/Admin/EventAdmin.php
@@ -239,7 +239,6 @@ class EventAdmin extends AbstractBaseAdmin
      */
     protected function configureListFields(ListMapper $listMapper): void
     {
-        unset($this->listModes['mosaic']);
         $listMapper
             ->add(
                 'begin',