+++ src/Admin/InvoiceAdmin.php
@@ -501,7 +501,6 @@ class InvoiceAdmin extends AbstractBaseAdmin
      */
     protected function configureListFields(ListMapper $listMapper): void
     {
-        unset($this->listModes['mosaic']);
         $listMapper
             ->add(
                 'id',