+++ src/Admin/InvoiceLineAdmin.php
@@ -147,7 +147,6 @@ class InvoiceLineAdmin extends AbstractBaseAdmin
      */
     protected function configureListFields(ListMapper $listMapper): void
     {
-        unset($this->listModes['mosaic']);
         $listMapper
             ->add(
                 'invoice',