+++ src/Admin/PersonAdmin.php
@@ -283,7 +283,6 @@ class PersonAdmin extends AbstractBaseAdmin
      */
     protected function configureListFields(ListMapper $listMapper): void
     {
-        unset($this->listModes['mosaic']);
         $listMapper
             ->add(
                 'name',