+++ src/Admin/ProvinceAdmin.php
@@ -111,7 +111,6 @@ class ProvinceAdmin extends AbstractBaseAdmin
      */
     protected function configureListFields(ListMapper $listMapper): void
     {
-        unset($this->listModes['mosaic']);
         $listMapper
             ->add(
                 'code',