+++ src/Admin/ServiceAdmin.php
@@ -127,7 +127,6 @@ class ServiceAdmin extends AbstractBaseAdmin
      */
     protected function configureListFields(ListMapper $listMapper): void
     {
-        unset($this->listModes['mosaic']);
         $listMapper
             ->add(
                 'position',