+++ src/Admin/TeacherAbsenceAdmin.php
@@ -127,7 +127,6 @@ class TeacherAbsenceAdmin extends AbstractBaseAdmin
      */
     protected function configureListFields(ListMapper $listMapper): void
     {
-        unset($this->listModes['mosaic']);
         $listMapper
             ->add(
                 'image',