+++ src/Admin/TeacherAdmin.php
@@ -154,7 +154,6 @@ class TeacherAdmin extends AbstractBaseAdmin
      */
     protected function configureListFields(ListMapper $listMapper): void
     {
-        unset($this->listModes['mosaic']);
         $listMapper
             ->add(
                 'position',