+++ src/Admin/UserAdmin.php
@@ -160,7 +160,6 @@ class UserAdmin extends ParentUserAdmin
      */
     protected function configureListFields(ListMapper $listMapper): void
     {
-        unset($this->listModes['mosaic']);
         $listMapper
             ->add(
                 'username',